backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.32
psycopg[binary]==3.3.1
python-dotenv==1.0.1
sentry-sdk==2.15.0
pydantic==1.10.17
httpx==0.27.2
pytest==8.3.3
pdf2image
Pillow
piexif
supabase
python-multipart